+++ c/frame/py_frame_init.cc
@@ -338,7 +338,7 @@ class FrameInitializationManager {
 
     void init_from_frame() {
       DataTable* srcdt = src.to_frame();
-      size_t ncols = static_cast<size_t>(srcdt->ncols);
+      size_t ncols = srcdt->ncols;
       check_names_count(ncols);
       if (stypes_arg || stype_arg) {
         // TODO: allow this use case
