+++ c/utils/parallel.h
@@ -147,9 +147,9 @@ class mapper_str2str : private ordered_job {
 
   public:
     mapper_str2str(StringColumn<T>* col, F _f)
-      : ordered_job(static_cast<size_t>(col->nrows)),
+      : ordered_job(col->nrows),
         inpcol(col),
-        outcol(static_cast<size_t>(col->nrows)),
+        outcol(col->nrows),
         f(_f) {}
     ~mapper_str2str() override = default;
